feat(docs): creating docs for reference for the addition of DEXs
summary: this commit features AI generated docs which outline the process for adding DEXs BalancerV2 and UniswapV3. Whilst these have been written in a sense which would support the live integration (as opposed to full redeployment) of the protocol, the likelihood is the protocol will be redeployed at some soon point in time. These docs nonetheless provide a succinct reference point for the deployment and maintenance of these DEXs.