feat(misc): creating updated workflow file for bot
summary: this commit features an updated file which runs our bot `local-monitor` on a cron cycle of 5 minutes. It has been implemented for the sake of testing and is expected to be updated in subsequent commits as we debug and optimise.

N.B. The cycle of once every 5 minutes is in place so as to ensure we have good time to observe execution sequences for the current contract implementation, where a default minimum of the sweet spot is set at 4. In production, this will be 1 (no caching at all and thus no requirement for bot) for nominally small trade volumes.